.pub-cache/hosted/pub.dartlang.org/html-0.15.0/test/data/tree-construction/tests12.dat

#data
<!DOCTYPE html><body><p>foo<math><mtext><i>baz</i></mtext><annotation-xml><svg><desc><b>eggs</b></desc><g><foreignObject><P>spam<TABLE><tr><td><img></td></table></foreignObject></g><g>quux</g></svg></annotation-xml></math>bar
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       "foo"
|       <math math>
|         <math mtext>
|           <i>
|             "baz"
|         <math annotation-xml>
|           <svg svg>
|             <svg desc>
|               <b>
|                 "eggs"
|             <svg g>
|               <svg foreignObject>
|                 <p>
|                   "spam"
|                 <table>
|                   <tbody>
|                     <tr>
|                       <td>
|                         <img>
|             <svg g>
|               "quux"
|       "bar"

#data
<!DOCTYPE html><body>foo<math><mtext><i>baz</i></mtext><annotation-xml><svg><desc><b>eggs</b></desc><g><foreignObject><P>spam<TABLE><tr><td><img></td></table></foreignObject></g><g>quux</g></svg></annotation-xml></math>bar
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "foo"
|     <math math>
|       <math mtext>
|         <i>
|           "baz"
|       <math annotation-xml>
|         <svg svg>
|           <svg desc>
|             <b>
|               "eggs"
|           <svg g>
|             <svg foreignObject>
|               <p>
|                 "spam"
|               <table>
|                 <tbody>
|                   <tr>
|                     <td>
|                       <img>
|           <svg g>
|             "quux"
|     "bar"
